RR-231 Moved all event form artifacts to client
