--- src/apps/events/client/Details/state.js
@@ -1,7 +1,0 @@
-export const TASK_GET_EVENT_DETAILS = 'TASK_GET_EVENT_DETAILS'
-
-export const ID = 'eventDetails'
-
-export const state2props = ({ ...state }) => ({
-  ...state[ID],
-})
