+++ src/client/modules/Events/EventForm/transformers.js
@@ -7,7 +7,7 @@ import {
   transformArrayOfUniqueOptions,
   transformOption,
   transformYesNoToBool,
-} from '../../../../client/transformers'
+} from '../../../transformers'
 
 const transformResponseToEventForm = (data) => {
   const {
